Fix unsoundness in copy_slot()
This fixes what appears to be 2 related issues in copy_slot().
First, because we tag the `atomic_value_slot`, this appears to change
the pointer that we `compare_and_set_*` with. Therefore, we could not
assume that the `primed_old_value`, which was previously obtained by
loading the `atomic_value_slot`, was still the same. This was the reason
that the non-weak compare_and_set_owned() failed according to the FIXME.

Diagrammatically speaking, we assumed that V' == V'_tagged and could
compare_and_swap() using either in the following state transitions:
(K, V) -> (K, V')
(K, V') -> (K, V'_tagged)
(K, V'_tagged) -> (K, X)

The solution to this was to not tag the `atomic_value_slot` and only tag
the `atomic_key_slot`, which we know will not be changed until it is
dropped. The `atomic_key_slot` is only dropped when the `MapInner` is
dropped, which cannot happen while we have the copy_slot() `&self`
parameter. This seems to inadvertently fix the second FIXME, which was
that the `!atomic_value_slot.is_tagged()` conditional would sometimes
segfault.