Fix unsoundness for promote()
Since promote() was called in the `if upper_bound == self.capacity()`
conditional, it can accidentally promote the map before it's ready to be
promoted. Thus, there is a need to introduce `try_promote()`, which
checks if all the slots have been copied before attempting to promote()
the map. To do this, the new variable `slots_copied` needed to be
introduced so that we can keep track of the number of slots already
copied at some point in time.